Initial textField generation for components

+++ Ohmwards/src/main/java/com/mycompany/ohmwards/Relay.java
@@ -13,17 +13,17 @@ import javax.swing.JLabel;
  * @author Cameron
  */
 public class Relay extends PowerSource {
-    private int voltsRequired;
+    private double voltsRequired;
     
     public Relay(Vector<Double> position, JButton btn) {
         super(9, ConnectionType.POWERSOURCE, position, btn, new JLabel("Output:"), new JLabel("Required Vols:"));
         this.voltsRequired = 12;
     }
     
-    public int getVoltsRequired() {
+    public double getVoltsRequired() {
         return voltsRequired;
     }
-    public void setVoltsRequired(int newRequired){
+    public void setVoltsRequired(double newRequired){
         voltsRequired = newRequired;
     }
 }
